requirements.txt
fastapi
uvicorn[standard]
line-bot-sdk==3.5.0
requests
google-cloud-pubsub